requirements.txt
h5py==3.7.0
keras==2.10.0
numpy==1.23.3
pandas==1.5.0
Pillow==9.2.0
streamlit==1.12.2
tensorflow-cpu==2.10.0


